Add RandomGraph
